+++ src/api-type-helpers.ts
@@ -80,6 +80,12 @@ export type ListFactMetricsResponse =
 export type GetFactMetricResponse =
   Paths["/fact-metrics/{id}"]["get"]["responses"][200]["content"]["application/json"];
 
+// Fact tables
+export type ListFactTablesResponse =
+  Paths["/fact-tables"]["get"]["responses"][200]["content"]["application/json"];
+export type GetFactTableResponse =
+  Paths["/fact-tables/{id}"]["get"]["responses"][200]["content"]["application/json"];
+
 // Stale features
 export type GetStaleFeatureResponse =
   Paths["/stale-features"]["get"]["responses"][200]["content"]["application/json"];